Move commands, change name and add aliases commands
