+++ src/App/YiiDevToolApplication.php
@@ -13,8 +13,6 @@ use Symfony\Component\Console\Input\InputDefinition;
 use Symfony\Component\Console\Input\InputOption;
 use Yiisoft\YiiDevTool\App\Command\Composer\ComposerFixDependenciesCommand;
 use Yiisoft\YiiDevTool\App\Command\Composer\UpdateCommand as ComposerUpdateCommand;
-use Yiisoft\YiiDevTool\App\Command\DisableCommand;
-use Yiisoft\YiiDevTool\App\Command\EnableCommand;
 use Yiisoft\YiiDevTool\App\Command\ExecCommand;
 use Yiisoft\YiiDevTool\App\Command\Git\CheckoutCommand;
 use Yiisoft\YiiDevTool\App\Command\Git\CloneCommand;
@@ -24,13 +22,17 @@ use Yiisoft\YiiDevTool\App\Command\Git\PushCommand;
 use Yiisoft\YiiDevTool\App\Command\Git\RequestPullCommand;
 use Yiisoft\YiiDevTool\App\Command\Git\StatusCommand;
 use Yiisoft\YiiDevTool\App\Command\Github\ForksRepositoriesCommand;
-use Yiisoft\YiiDevTool\App\Command\Github\SyncUpstreamRepositoriesCommand;
 use Yiisoft\YiiDevTool\App\Command\Github\ProtectBranchCommand;
 use Yiisoft\YiiDevTool\App\Command\Github\SettingsCommand;
-use Yiisoft\YiiDevTool\App\Command\InstallCommand;
+use Yiisoft\YiiDevTool\App\Command\Github\SyncUpstreamRepositoriesCommand;
 use Yiisoft\YiiDevTool\App\Command\LintCommand;
 use Yiisoft\YiiDevTool\App\Command\Packages\AddCommand;
+use Yiisoft\YiiDevTool\App\Command\Packages\DisableCommand;
+use Yiisoft\YiiDevTool\App\Command\Packages\EnableCommand;
+use Yiisoft\YiiDevTool\App\Command\Packages\InstallCommand;
+use Yiisoft\YiiDevTool\App\Command\Packages\ListCommand;
 use Yiisoft\YiiDevTool\App\Command\Packages\RemoveCommand;
+use Yiisoft\YiiDevTool\App\Command\Packages\UpdateCommand;
 use Yiisoft\YiiDevTool\App\Command\Release\MakeCommand;
 use Yiisoft\YiiDevTool\App\Command\Release\WhatCommand;
 use Yiisoft\YiiDevTool\App\Command\Replicate\ReplicateComposerConfigCommand;
@@ -39,7 +41,6 @@ use Yiisoft\YiiDevTool\App\Command\Replicate\ReplicateFilesCommand;
 use Yiisoft\YiiDevTool\App\Command\Stats\ContributorsCommand;
 use Yiisoft\YiiDevTool\App\Command\TestCommand;
 use Yiisoft\YiiDevTool\App\Command\Tool\Init;
-use Yiisoft\YiiDevTool\App\Command\UpdateCommand;
 use Yiisoft\YiiDevTool\App\Component\Config;
 
 final class YiiDevToolApplication extends Application
